Massive GUI Update
Added DataGridView above the previously in place ListBoxes. See
screenshot in readme. The data for the grid is gathered from the XML
library file. MultiSort was implemented to sort the grid by multiple
columns depending on which one is selected. This is for ease of use,
mainly so that songs still show up by album, and in track-number-order
when helpful.

Slight XML implementation change. Single digit track numbers now have a
leading '0' for string comparison purposes.